+++ src/main/java/io/github/forgestove/create_cyber_goggles/CCGConfig.java
@@ -22,6 +22,7 @@ public class CCGConfig implements ConfigData {
 		@Tooltip public boolean renderExtraItems = true;
 		@Tooltip public boolean enableKineticEffect = true;
 		@Tooltip public boolean preciseNumbers = true;
+		@Tooltip public boolean disableScreenGoggles = true;
 		public static class GameMode {
 			@Tooltip public boolean enableInSurvival = true;
 			@Tooltip public boolean enableInCreative = true;